+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog records changes to stable releases since 1.50.2. "TBA" changes he
 - fix: race when stopping on node-internals exceptions ([vscode#204581](https://github.com/microsoft/vscode/issues/204581))
 - fix: off-by-one error leading to invalid renames ([#1948](https://github.com/microsoft/vscode-js-debug/issues/1948))
 - fix: apply sourceMapPathOverrides to sourceURL scripts ([vscode#204784](https://github.com/microsoft/vscode/issues/204784))
+- fix: attempt both ipv4 and ipv6 loopbacks for DWARF symbols
 
 ## v1.87 (February 2024)